// src/services/product/models/BrandDTO.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type BrandDTO = {
    /**
     * brand identifier
     */
    id?: string;
    /**
     * brand name
     */
    name?: string;
    /**
     * brand description
     */
    description?: string;
};